+++ Models/User.cs
@@ -99,6 +99,12 @@ namespace Models
         /// </summary>
         public List<UserTask> UserTasks { get; set; }
 
+        public bool IsDataOfficerForInstitution(int institutionId)
+        {
+            return InstitutionId != null &&
+                   InstitutionId.Value == institutionId &&
+                   Role.Name == Defaults.Defaults.Roles.DataOfficer;
+        }
     }
 
 }