+++ Repositories/ProjectRepository.cs
@@ -44,7 +44,7 @@ namespace Repositories
         /// <param name="orderByAsc"></param>
         /// <param name="highlighted"></param>
         /// <returns>List of projects</returns>
-        Task<List<Project>> GetAllWithUsersAndCollaboratorsAsync(
+        Task<List<Project>> GetAllWithUsersCollaboratorsAndInstitutionsAsync(
             int? skip = null,
             int? take = null,
             Expression<Func<Project, object>> orderBy = null,
@@ -94,7 +94,9 @@ namespace Repositories
         /// <returns>
         ///     This method returns possibly redacted Project object with user and collaborators.
         /// </returns>
-        Task<Project> FindWithUserAndCollaboratorsAsync(int id);
+        Task<Project> FindWithUserCollaboratorsAndInstitutionsAsync(int id);
+
+        Task<bool> ProjectExistsAsync(int id);
 
         /// <summary>
         ///     Get the user projects.
@@ -153,7 +155,7 @@ namespace Repositories
         /// <param name="orderByAsc">The order by asc parameters represents the order direction (True: asc, False: desc)</param>
         /// <param name="highlighted">The highlighted parameter represents the whether to filter highlighted projects.</param>
         /// <returns>This method returns a list of projects filtered by the specified parameters.</returns>
-        public virtual async Task<List<Project>> GetAllWithUsersAndCollaboratorsAsync(
+        public virtual async Task<List<Project>> GetAllWithUsersCollaboratorsAndInstitutionsAsync(
             int? skip = null,
             int? take = null,
             Expression<Func<Project, object>> orderBy = null,
@@ -167,7 +169,8 @@ namespace Repositories
                                                     .Include(p => p.CallToAction)
                                                     .Include( p => p.Collaborators )
                                                     .Include( p => p.User )
-                                                    .Include( p => p.Likes );
+                                                    .Include( p => p.Likes )
+                                                    .Include(p => p.LinkedInstitutions);
 
             queryableProjects = ApplyFilters(queryableProjects, skip, take, orderBy, orderByAsc, highlighted);
 
@@ -237,7 +240,7 @@ namespace Repositories
         /// <returns>
         ///     This method returns possibly redacted Project object with user and collaborators.
         /// </returns>
-        public async Task<Project> FindWithUserAndCollaboratorsAsync(int id)
+        public async Task<Project> FindWithUserCollaboratorsAndInstitutionsAsync(int id)
         {
             Project project = await GetDbSet<Project>()
                                     .Include(p => p.User)
@@ -253,6 +256,11 @@ namespace Repositories
                 project.Likes = await GetDbSet<ProjectLike>()
                                       .Where(p => p.LikedProject.Id == project.Id)
                                       .ToListAsync();
+
+                project.LinkedInstitutions = await GetDbSet<ProjectInstitution>()
+                                                    .Include(p => p.Institution)
+                                                    .Where(p => p.ProjectId == project.Id)
+                                                    .ToListAsync();
             }
 
             return RedactUser(project);
@@ -461,6 +469,10 @@ namespace Repositories
             return projectsToReturn;
         }
 
+        public Task<bool> ProjectExistsAsync(int id)
+        {
+            return GetDbSet<Project>().AnyAsync(i => i.Id == id);
+        }
     }
 
 }